Updated MCP Mappings

+++ src/main/kotlin/org/jglrxavpok/moarboats/client/renders/SonarModuleRenderer.kt
@@ -53,7 +53,7 @@ object SonarModuleRenderer : BoatModuleRenderer() {
 
                         val angle = Math.atan2(gradientVal.y.toDouble(), gradientVal.x.toDouble()).toDegrees()
                         GlStateManager.rotatef(angle.toFloat(), 0f, 1f, 0f)
-                        GlStateManager.scalef(0.1f, 0.1f, gradientVal.length() * 0.1f)
+                        GlStateManager.scalef(0.1f, 0.1f, gradientVal.length().toFloat() * 0.1f)
                         if(potentialState.block is BlockLiquid) {
                             Minecraft.getInstance().blockRendererDispatcher.renderBlockBrightness(Blocks.EMERALD_BLOCK.defaultState, boat.brightness)
                         } else {
